requirements.txt
brotlipy==0.7.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.1
cycler==0.11.0
fonttools==4.25.0
future==0.18.2
idna==3.4
kiwisolver==1.4.2
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pip==22.1.2
ply==3.11
pycocotools==2.0.5
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
setuptools==63.4.1
sip==6.6.2
six==1.16.0
toml==0.10.2
torch==1.13.0.dev20220928
torch-lr-finder==0.2.1
torchaudio==0.13.0.dev20220928
torchvision==0.14.0.dev20220928
tornado==6.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
utils==1.0.1
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
